scripts/requirements.txt
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.1.0
numpy==1.24.3
img2pdf==0.5.1
rembg==2.0.50
requests==2.31.0
